Move MaxFileSizeAttribute to Validators namespace and delete old implementation
