+++ aspnet-core/src/toyiyo.todo.Application/Jobs/Dto/JobImageCreateInputDto.cs
@@ -1,6 +1,7 @@
 using System;
 using System.ComponentModel.DataAnnotations;
 using Microsoft.AspNetCore.Http;
+using toyiyo.todo.Jobs.Validators;
 
 namespace toyiyo.todo.Jobs
 {
