--- aspnet-core/src/toyiyo.todo.Application/Jobs/Dto/MaxFileSizeAttribute.cs
@@ -1,26 +1,0 @@
-using System.ComponentModel.DataAnnotations;
-using Microsoft.AspNetCore.Http;
-
-public class MaxFileSizeAttribute : ValidationAttribute
-{
-    private readonly int _maxFileSize;
-
-    public MaxFileSizeAttribute(int maxFileSize)
-    {
-        _maxFileSize = maxFileSize;
-    }
-
-    protected override ValidationResult IsValid(object value, ValidationContext validationContext)
-    {
-        var file = value as IFormFile;
-        if (file != null)
-        {
-            if (file.Length > _maxFileSize)
-            {
-                return new ValidationResult($"Maximum allowed file size is {_maxFileSize} bytes.");
-            }
-        }
-
-        return ValidationResult.Success;
-    }
-}
